# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Oralia API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guidance/today:
    get:
      operationId: getTodayGuidance
      tags: [guidance]
      summary: Get today's cached guidance
      responses:
        "200":
          description: Today's guidance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyGuidance"
        "404":
          description: No guidance for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List location profiles
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationProfile"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a location profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationProfile"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a location profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationProfile"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationBody"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationProfile"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Location deleted
  /locations/{id}/strategy:
    post:
      operationId: generateLocationStrategy
      tags: [locations]
      summary: Generate AI location strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location strategy generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationProfile"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /library:
    get:
      operationId: listLibraryEntries
      tags: [library]
      summary: List library entries
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of library entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryEntry"
...
  /library/{id}:
    get:
      operationId: getLibraryEntry
      tags: [library]
      summary: Get a library entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Library entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DailyGuidance:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        theme:
          type: string
        bestUse:
          type: string
        avoid:
          type: string
        career:
          type: string
        relationship:
          type: string
        body:
          type: string
        chakra:
          type: string
        moon:
          type: string
        goalNudge:
          type: string
        action:
          type: string
        journalPrompt:
          type: string
        ritual:
          type: string
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - date
        - theme
        - bestUse
        - avoid
        - career
        - relationship
        - body
        - chakra
        - moon
        - goalNudge
        - action
        - journalPrompt
        - ritual
        - isAiGenerated
...
    LocationProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        city:
          type: string
        country:
          type: string
        locationType:
          type: string
        locationGoal:
          type: string
        bestUse:
          type: ["string", "null"]
        whatToDo:
          type: ["string", "null"]
        whatNotToDo:
          type: ["string", "null"]
        bestTimingStyle:
          type: ["string", "null"]
        recommendedPurpose:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - city
        - country
        - locationType
        - locationGoal
    CreateLocationBody:
      type: object
      properties:
        city:
          type: string
        country:
          type: string
        locationType:
          type: string
        locationGoal:
          type: string
      required:
        - city
        - country
        - locationType
        - locationGoal
    UpdateLocationBody:
      type: object
      properties:
        city:
          type: string
        country:
          type: string
        locationType:
          type: string
        locationGoal:
          type: string
    LibraryEntry:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - title
        - content
...